move list_or_dict
