+++ qiskit_optimization/minimum_eigensolvers/minimum_eigensolver.py
@@ -20,7 +20,7 @@ from typing import Any
 from qiskit.quantum_info.operators.base_operator import BaseOperator
 
 from ..algorithm_result import AlgorithmResult
-from ..list_or_dict import ListOrDict
+from .list_or_dict import ListOrDict
 
 
 class MinimumEigensolver(ABC):
