+++ qiskit_optimization/minimum_eigensolvers/sampling_mes.py
@@ -22,7 +22,7 @@ from qiskit.quantum_info.operators.base_operator import BaseOperator
 from qiskit.result import QuasiDistribution
 
 from ..algorithm_result import AlgorithmResult
-from ..list_or_dict import ListOrDict
+from .list_or_dict import ListOrDict
 
 
 class SamplingMinimumEigensolver(ABC):
