+++ qiskit_optimization/minimum_eigensolvers/sampling_vqe.py
@@ -27,7 +27,6 @@ from qiskit.quantum_info.operators.base_operator import BaseOperator
 from qiskit.result import QuasiDistribution
 
 from ..exceptions import AlgorithmError
-from ..list_or_dict import ListOrDict
 from ..minimum_eigensolvers.sampling_mes import (
     SamplingMinimumEigensolver,
     SamplingMinimumEigensolverResult,
@@ -39,6 +38,7 @@ from ..utils.primitives import _init_observable
 # private function as we expect this to be updated in the next released
 from ..utils.set_batching import _set_default_batchsize
 from .diagonal_estimator import _DiagonalEstimator
+from .list_or_dict import ListOrDict
 from .observables_evaluator import estimate_observables
 from .variational_algorithm import VariationalAlgorithm, VariationalResult
