+++ qiskit_optimization/minimum_eigensolvers/vqe.py
@@ -26,13 +26,13 @@ from qiskit.primitives import BaseEstimatorV1, BaseEstimatorV2
 from qiskit.quantum_info.operators.base_operator import BaseOperator
 
 from ..exceptions import AlgorithmError
-from ..list_or_dict import ListOrDict
 from ..optimizers import Minimizer, Optimizer, OptimizerResult
 from ..utils import validate_bounds, validate_initial_point
 from ..utils.primitives import _init_observable
 
 # private function as we expect this to be updated in the next released
 from ..utils.set_batching import _set_default_batchsize
+from .list_or_dict import ListOrDict
 from .minimum_eigensolver import MinimumEigensolver, MinimumEigensolverResult
 from .observables_evaluator import estimate_observables
 from .variational_algorithm import VariationalAlgorithm, VariationalResult
